+++ kobay.com/src/main/webapp/WEB-INF/view/member/editProfile.jsp
@@ -8,23 +8,12 @@
 <!DOCTYPE html>
 <html lang="utf-8">
   <head>
-	<!-- <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="">
-    <meta name="author" content="">
-
-    <title>Kobay 로그인 및 회원가입</title>
-
-    Bootstrap core CSS
-    <link rel="stylesheet" type="text/css" href="../../../vendor/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-
-    Custom styles for this template
-    <link rel="stylesheet" href="../../../css/kobay.css" > -->
 	<link rel="stylesheet" type="text/css" href="../../../css/member.css">
   </head>
 <script type="text/javascript">
+/* edit 폼에 있는 인풋 id 또는 name */
 var editList = ["memberId","memberPwd","memberName","memberPhone"];
+/*  */
 /* 회원정보 수정 */
 function member_Update() {
 	var frm = $("#editForm").serialize();
@@ -36,13 +25,13 @@ function member_Update() {
 		dataType:"json",
 		async: false,
 		success:function(data) {
-			if(data.result == "ok") {
+			if(data.result == "ok") { /* 변경 성공 */
 				alert("수정되었습니다.\n다시 로그인해주세요.");
 				location.href="<c:url value='/logout' />";
 			}
 			else{
 				alert("다시 확인해주세요.");
-				for(var i=0; i<data.errors.length; i++) {
+				for(var i=0; i<data.errors.length; i++) { /* 라벨 옆에 v 또는 x 표시 */
 					$("#editForm i[id='icon_"+data.errors[i].field+"']").attr("class","fa fa-ban");
 					$("#editForm i[id='icon_"+data.errors[i].field+"']").css({
 						 "color":"#FF0000"
@@ -69,31 +58,21 @@ function member_CheckForm(va) {
 			dataType:"json",
 			success:function(data) {
 				if(data.result == "ok") 
-				{
-					if(data.errors.length > 0) 
-					{
-						for(var i=0; i<data.errors.length; i++) 
-						{
-							if(data.errors[i].field == va)
-							{
-								$("#editForm i[id='icon_"+va+"']").attr("class","fa fa-ban");
-								$("#editForm i[id='icon_"+va+"']").css({"color":"#FF0000"});
-								return;
-							}							
-							else 
-							{
-								$("#editForm i[id='icon_"+va+"']").attr("class","fa fa-check");
-								$("#editForm i[id='icon_"+va+"']").css({"color":"#1DDB16"});
-							}
-						}
-					}
+				{	/* 전부 사용 가능 */
+					$("#editForm i[id='icon_"+va+"']").attr("class","fa fa-check");
+					$("#editForm i[id='icon_"+va+"']").css({"color":"#1DDB16"});
 				}
-				else 
+				else
 				{
 					for(var i=0; i<editList.length; i++)
-					{
+					{ /* 일단 전부 v체크 후 밑에서 x로 바꿈 */
 						$("#editForm i[id='icon_"+editList[i]+"']").attr("class","fa fa-check");
 						$("#editForm i[id='icon_"+editList[i]+"']").css({"color":"#1DDB16"});
+					}
+					for(var i=0; i<data.errors.length; i++) 
+					{ /* 유효성검사에 따른 x 체크 */
+						$("#editForm i[id='icon_"+data.errors[i].field+"']").attr("class","fa fa-ban");
+						$("#editForm i[id='icon_"+data.errors[i].field+"']").css({"color":"#FF0000"});
 					}
 				} 
 				return false;
@@ -103,10 +82,9 @@ function member_CheckForm(va) {
 			}
 		}); 
 }
+/*  */
 </script>
-
 <body>
-    
     <!-- Page Content -->
     <div class="container">
 		<!-- 수정 부분 -->
@@ -117,7 +95,7 @@ function member_CheckForm(va) {
 		<form:form commandName="editForm" name="editForm" class="form-horizontal" onsubmit="return false;">					
 			<div class="form-group">
 				<form:label path="memberId" class="col-sm-2 control-label">
-					이메일<i class="fa fa-check" id="icon_memberId" aria-hidden="true" style="color:#1DDB16;"></i>
+					이메일
 				</form:label>
 				<form:input path="memberId" class="form-control" value="${vo.memberId}" onchange="member_CheckForm('memberId')" readonly="true" />
 			 </div>
